requirements.txt
Django==5.0.7
# psycopg2-binary==2.9.9  # Optional for PostgreSQL; disabled for local SQLite migrations
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0